+++ dspace-server-webapp/src/main/java/org/dspace/app/rest/security/clarin/ClarinShibbolethLoginFilter.java
@@ -307,7 +307,12 @@ public class ClarinShibbolethLoginFilter extends StatelessLoginFilter {
 
         if (StringUtils.equalsAnyIgnoreCase(redirectHostName, allowedHostNames.toArray(new String[0]))) {
             log.debug("Shibboleth redirecting to " + redirectUrl);
-            response.sendRedirect(redirectUrl);
+            // Encode the UTF-8 characters from redirect URL to UTF-8, to ensure it's properly encoded for the browser
+            String encodedRedirectUrl = org.dspace.app.rest.utils.Utils.encodeNonAsciiCharacters(redirectUrl);
+            if (StringUtils.isEmpty(encodedRedirectUrl)) {
+                log.error("Invalid Encoded Shibboleth redirectURL=" + redirectUrl + ". URL is empty!");
+            }
+            response.sendRedirect(encodedRedirectUrl);
         } else {
             log.error("Invalid Shibboleth redirectURL=" + redirectUrl +
                     ". URL doesn't match hostname of server or UI!");
